requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
conda==4.6.14
cryptography==2.4.2
Django==2.2.1
idna==2.8
mysql-connector-python==8.0.16
protobuf==3.7.1
pycosat==0.6.3
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==18.0.0
PySocks==1.6.8
pytz==2019.1
requests==2.21.0
ruamel-yaml==0.15.46
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
